[Fusilli] Add clang-tidy checks to CI (#2607)
Adds `FUSILLI_ENABLE_CLANG_TIDY` option to cmake and enables it on CI.
Running `clang-tidy` through CLI/cmake performs more checks than running
in an editor (uses the clang-tidy integrated into clangd). I disabled
some checks that we should reevaluate to see if they should be kept or
not.

For example, I disabled `misc-const-correctness` since it was not
showing up as an error in my editor but was erroring out the build.
Fixing this would mean editing large portions of almost every file.
Then, once added, it would probably cause a headache for devs due to
lack of editor support.

I'm suggesting to first enable clang-tidy on CI (with a few checks
disabled) so that new PRs can be analyzed against the current checks.
And then over time we can adjust the checks as necessary.


Fixes all instances of:
- `misc-include-cleaner`
- `readability-identifier-naming`
- `performance-unnecessary-copy-initialization`
- `misc-use-internal-linkage`

---------

Signed-off-by: Ian Wood <ianwood@u.northwestern.edu>