[Fusilli] Cleanup LINALG-CHECK from certain lit tests (#2646)
Context:
https://github.com/nod-ai/shark-ai/pull/2629#discussion_r2495582000

> For this case though, I feel we can just drop the LINALG-CHECKS
entirely, because:
> 1. The IR is sub-optimal and will go through significant changes with
the torch-mlir revamp of backward conv lowering that you and others are
working on).
> 2. Tests on linalg transformations would normally live alongside the
sources for those linalg transformations - so torch-mlir or IREE (we
don't do any of that in fusilli as we stop at emitting torch dialect).

> We made an exception to 2 for some lit tests and added the linalg
checks to verify certain assumptions we make in fusilli around layout
transformations - like emitting transposes for NHWC -> NCHW don't incur
an actual cost and are absorbed by transpose propagation/eliminations in
the compiler (Torch->Linalg). Sorry for the confusion.