Add flake detection workflow to CI (#35)
We recently encountered a GPU stall / segfault related flake that took a
while to bisect into. This should help catch such flake issues from
being introduced, or at least triaged quicker without needing to bisect.

I haven't thought of all the trade-offs in making this gate merges but I
plan to until we see a good reason not to.

---------

Signed-off-by: Sambhav Jain <sambhav@alumni.stanford.edu>